requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.2,<3.16
psycopg2>=2.9.10,<2.10
drf-spectacular>=0.28.0,<0.29.0
djangorestframework-simplejwt>=5.5.0,<5.6
django-filter>=25.1,<26.0
gunicorn>=21.2.0
django-cors-headers>=4.3.1,<4.4.0
pillow>=11.0.0,<12.0.0
uwsgi>=2.0.20,<2.1
dos2unix
